Extract database module.
+ Relates to issue #53.
