+++ app/src/main/java/nerd/tuxmobil/fahrplan/congress/serialization/FahrplanParser.java
@@ -16,16 +16,16 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
+import info.metadude.android.eventfahrplan.database.contract.FahrplanContract.LecturesTable;
+import info.metadude.android.eventfahrplan.database.contract.FahrplanContract.LecturesTable.Columns;
+import info.metadude.android.eventfahrplan.database.contract.FahrplanContract.LecturesTable.Values;
+import info.metadude.android.eventfahrplan.database.contract.FahrplanContract.MetasTable;
+import info.metadude.android.eventfahrplan.database.sqliteopenhelper.LecturesDBOpenHelper;
+import info.metadude.android.eventfahrplan.database.sqliteopenhelper.MetaDBOpenHelper;
 import nerd.tuxmobil.fahrplan.congress.MyApp;
 import nerd.tuxmobil.fahrplan.congress.contract.BundleKeys;
 import nerd.tuxmobil.fahrplan.congress.models.Lecture;
 import nerd.tuxmobil.fahrplan.congress.models.MetaInfo;
-import nerd.tuxmobil.fahrplan.congress.persistence.FahrplanContract.LecturesTable;
-import nerd.tuxmobil.fahrplan.congress.persistence.FahrplanContract.LecturesTable.Columns;
-import nerd.tuxmobil.fahrplan.congress.persistence.FahrplanContract.LecturesTable.Values;
-import nerd.tuxmobil.fahrplan.congress.persistence.FahrplanContract.MetasTable;
-import nerd.tuxmobil.fahrplan.congress.persistence.LecturesDBOpenHelper;
-import nerd.tuxmobil.fahrplan.congress.persistence.MetaDBOpenHelper;
 import nerd.tuxmobil.fahrplan.congress.serialization.exceptions.MissingXmlAttributeException;
 import nerd.tuxmobil.fahrplan.congress.utils.DateHelper;
 import nerd.tuxmobil.fahrplan.congress.utils.FahrplanMisc;
