+++ app/src/main/java/nerd/tuxmobil/fahrplan/congress/utils/FahrplanMisc.java
@@ -26,6 +26,14 @@ import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 
+import info.metadude.android.eventfahrplan.database.contract.FahrplanContract.AlarmsTable;
+import info.metadude.android.eventfahrplan.database.contract.FahrplanContract.HighlightsTable;
+import info.metadude.android.eventfahrplan.database.contract.FahrplanContract.LecturesTable;
+import info.metadude.android.eventfahrplan.database.contract.FahrplanContract.MetasTable;
+import info.metadude.android.eventfahrplan.database.sqliteopenhelper.AlarmsDBOpenHelper;
+import info.metadude.android.eventfahrplan.database.sqliteopenhelper.HighlightDBOpenHelper;
+import info.metadude.android.eventfahrplan.database.sqliteopenhelper.LecturesDBOpenHelper;
+import info.metadude.android.eventfahrplan.database.sqliteopenhelper.MetaDBOpenHelper;
 import nerd.tuxmobil.fahrplan.congress.BuildConfig;
 import nerd.tuxmobil.fahrplan.congress.MyApp;
 import nerd.tuxmobil.fahrplan.congress.R;
@@ -36,14 +44,6 @@ import nerd.tuxmobil.fahrplan.congress.extensions.Contexts;
 import nerd.tuxmobil.fahrplan.congress.models.DateInfo;
 import nerd.tuxmobil.fahrplan.congress.models.DateInfos;
 import nerd.tuxmobil.fahrplan.congress.models.Lecture;
-import nerd.tuxmobil.fahrplan.congress.persistence.AlarmsDBOpenHelper;
-import nerd.tuxmobil.fahrplan.congress.persistence.FahrplanContract.AlarmsTable;
-import nerd.tuxmobil.fahrplan.congress.persistence.FahrplanContract.HighlightsTable;
-import nerd.tuxmobil.fahrplan.congress.persistence.FahrplanContract.LecturesTable;
-import nerd.tuxmobil.fahrplan.congress.persistence.FahrplanContract.MetasTable;
-import nerd.tuxmobil.fahrplan.congress.persistence.HighlightDBOpenHelper;
-import nerd.tuxmobil.fahrplan.congress.persistence.LecturesDBOpenHelper;
-import nerd.tuxmobil.fahrplan.congress.persistence.MetaDBOpenHelper;
 import nerd.tuxmobil.fahrplan.congress.wiki.WikiEventUtils;
 
 import static nerd.tuxmobil.fahrplan.congress.BuildConfig.COMPOSE_EVENT_URL_FROM_SLUG;
