+++ database/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="info.metadude.android.eventfahrplan.database" />
